🏇 improve response time
PR-URL: https://github.com/johnmuhl/electron-crash-report-server/pull/37
